fix: replace two-column layout with single-column PanelView cards
The previous two-column layout (StickyHeaderBar + LiveOutputPanel +
ConfigSidebar) was inconsistent with other pages and caused poor space
distribution. Switch to standard scrollable PageView with PanelView
cards matching the app's existing design pattern.
